requirements.txt
google_cloud_aiplatform
matplotlib
numpy
protobuf
python-dotenv
transformers
